combine abstract_plane_material_factory and plane_material_factory
